Massive refactor of python bindings
